% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_EC}
\alias{DK_EC}
\title{Electrical conductivity for five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_EC}}{

List of wrapped SpatRaster objects with one layer each.
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2020.114852}
}
\usage{
DK_EC
}
\description{
Raster layers showing the electrical conductivity for the topsoil (0 - 25 cm) in the five research fields described in Møller et al. (2020). Depth-specific values based on an inversion of data from a DUALEM21 sensor.
}
\examples{
list_unwrap(
  DK_EC,
  "EPSG:25832"
  )
}
\keyword{datasets}
