% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_observations}
\alias{DK_observations}
\title{Soil observations from five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_observations}}{

List of wrapped SpatVector point datasets.
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2020.114852}
}
\usage{
DK_observations
}
\description{
Points with measured soil properties from the five research fields described in Møller et al. (2020).
}
\examples{
list_unwrap(
  DK_observations,
  "EPSG:25832"
  )
}
\keyword{datasets}
