% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_RGB}
\alias{DK_RGB}
\title{Orthophotos for five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_RGB}}{

List of wrapped SpatRaster objects with three layers each.
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2020.114852}
}
\usage{
DK_RGB
}
\description{
Raster datasets showing RGB orthophotos for of the five research fields described in Møller et al. (2020). Based on averages from several images.
}
\examples{
list_unwrap(
  DK_RGB,
  "EPSG:25832"
  )
}
\keyword{datasets}
