% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_soilgrids}
\alias{DK_soilgrids}
\title{250-m soil maps for five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_soilgrids}}{

Wrapped SpatRaster dataset.
}
}
\source{
\url{https://soilgrids.org/}
}
\usage{
DK_soilgrids
}
\description{
Raster dataset showing soil maps from SoilGrids250m 2.0 for the five research fields described in Møller et al. (2020). Cropped to the areas surrounding the fields.
}
\examples{
library(terra)
DK_soilgrids <- unwrap(DK_soilgrids)
crs(DK_soilgrids) <- "EPSG:4326"
}
\keyword{datasets}
