% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_soilgrids_unc}
\alias{DK_soilgrids_unc}
\title{Uncertainties for 250-m soil maps for five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_soilgrids_unc}}{

Wrapped SpatRaster dataset.
}
}
\source{
\url{https://soilgrids.org/}
}
\usage{
DK_soilgrids_unc
}
\description{
Raster dataset showing uncertainties for soil maps from SoilGrids250m 2.0 for the five research fields described in Møller et al. (2020). The maps provide uncertainties as a 90\% prediction interval. Cropped to the areas surrounding the fields.
}
\examples{
library(terra)
DK_soilgrids_unc <- unwrap(DK_soilgrids_unc)
crs(DK_soilgrids_unc) <- "EPSG:4326"
}
\keyword{datasets}
