% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_unwrap.R
\name{list_unwrap}
\alias{list_unwrap}
\title{Unwrap a list of spatial objects}
\usage{
list_unwrap(x = NULL, crs_new = NULL)
}
\arguments{
\item{x}{A list of wrapped spatial objects created by \code{terra::wrap}}

\item{crs_new}{CRS to be assigned to the objects in the list (optional).}
}
\value{
A list of spatial objects
}
\description{
Unwrap a list of spatial objects
}
\examples{
list_unwrap(DK_observations, "EPSG:25832")
}
