% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_downscaler.R
\name{make_downscaler}
\alias{make_downscaler}
\title{Train a downscaler model}
\usage{
make_downscaler(
  obs = NULL,
  targ_name = NULL,
  cov = NULL,
  input = NULL,
  input_unc = NULL,
  model_type = "lm",
  make_maps = TRUE,
  unc_factor = 1,
  flatten_input = TRUE,
  input_as_cov = FALSE,
  scale_cov = c("no", "by_input", "by_SD"),
  scale_obs = FALSE,
  center_cov = TRUE,
  center_obs = TRUE,
  plot_results = FALSE,
  keep_obs = FALSE,
  keep_cov = FALSE,
  keep_models = TRUE
)
}
\arguments{
\item{obs}{List of spatvector points with observations.}

\item{targ_name}{Name of target variable.}

\item{cov}{List of SpatRaster objects containing the covariates for prediction.}

\item{input}{SpatRaster or list of SpatRaster objects containing the coarse resolution input map.}

\item{input_unc}{SpatRaster containing the uncertainty for the coarse-resolution input map.}

\item{model_type}{Model type passed to train. Ideally, I should use do.call, so users can pass a list of arguments to the training.}

\item{make_maps}{Should the function produce maps for the input sites when run?}

\item{unc_factor}{Multiplication factor for the uncertainties. Not implemented.}

\item{flatten_input}{Should the input be transformed to a mean value for each site?}

\item{input_as_cov}{Transform the input map to a covariate for mapping?}

\item{scale_cov}{Should the function scale the covariates relative to the input maps or their own standard deviations?  Default is no scaling.}

\item{scale_obs}{Should the function scale the observations relative to the input maps? Back-transformation is automatic for predictions. Default is no scaling.}

\item{center_cov}{Should the function center the covariates?}

\item{center_obs}{Should the function center the observations? Back-transformation is automatic for predictions.}

\item{plot_results}{Should the output contain a plot with the results?}

\item{keep_obs}{Append the modified observations to the output object?}

\item{keep_cov}{Append the covariates to the output object?}

\item{keep_models}{Keep the models for mapping as part of the output object?}
}
\value{
A downscaler model
}
\description{
Train a downscaler model
}
\examples{
library(magrittr)
library(terra)

data(DK_observations)
data(DK_soilgrids)
data(DK_EC)

my_obs <- list_unwrap(DK_observations, "EPSG:25832")

my_input <- DK_soilgrids \%>\% unwrap() \%>\% subset(1)
crs(my_input) <- "EPSG:4326"

my_cov <- list_unwrap(DK_EC, "EPSG:25832")

ds <- make_downscaler(
obs           = my_obs,
targ_name     = "clay",
input         = my_input,
make_maps     = FALSE,
cov           = my_cov
)

ds

}
