% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_fields}
\alias{DK_fields}
\title{Five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_fields}}{

A wrapped SpatVector polygons dataset with five geometries.
\describe{
\item{Site}{The name of the research field.}
...
}
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2020.114852}
}
\usage{
DK_fields
}
\description{
Polygon field extents for the research fields described in Møller et al. (2020)
}
\examples{
library(terra)
my_fields <- unwrap(DK_fields)
crs(my_fields) <- "EPSG:25832"
}
\keyword{datasets}
