% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{DK_observations}
\alias{DK_observations}
\title{Soil observations for five agricultural fields in Denmark.}
\format{
\subsection{\code{DK_observations}}{

List of wrapped spatvectors. Unwrap using DK_observations \%>\% lapply(function(x) terra::vect(x))
}
}
\source{
\url{https://doi.org/10.1016/j.geoderma.2020.114852}
}
\usage{
DK_observations
}
\description{
Field extents for the research fields described in Møller et al. (2020)
}
\keyword{datasets}
